[Site Isolation] Reduce the number of launched WebContent process in testing https://bugs.webkit.org/show_bug.cgi?id=299766 rdar://161570096
Reviewed by NOBODY (OOPS!).

When running layout tests in Site Isolation mode, we are not reusing the process used to load about:blank by WebKitTestRunner.
We should be able to reuse this process, since a process only used to load about:blank is not added to the process cache, so
there should not be another page using this process.

This change reduces the number of WebContent processes launched in layout test in Site Isolation mode with 1 per test, and
brings the total number down to 3 from 4. This is the lowest possible number, as long as we are not utilizing the process
cache in layout tests.

* Source/WebKit/UIProcess/WebProcessPool.cpp:
(WebKit::WebProcessPool::processForNavigationInternal):