The --site-isolation option in run-webkit-tests should only enable Site Isolation https://bugs.webkit.org/show_bug.cgi?id=300993 rdar://162874305
Reviewed by NOBODY (OOPS!).

Currently, this option is enabling tests to be run in a cross origin iframe, with and without Site Isolation
enabled. It may be clearer to rename the option to --enable-site-isolation and have it only enable Site
Isolation. We can instead use the --load-in-cross-origin-iframe option to run the tests in a cross origin
iframe, with Site Isolation on and off.

* Tools/Scripts/webkitpy/layout_tests/controllers/single_test_runner.py:
(SingleTestRunner):
* Tools/Scripts/webkitpy/layout_tests/run_webkit_tests.py:
(parse_args):
(_set_up_derived_options):